+++ cookbooks/chef-drupal/CHANGELOG.md
@@ -1,6 +1,14 @@
 chef-drupal CHANGELOG
 =====================
 
+0.1.2
+-----
+- nollieheel - Now with own definition for creating settings file
+
+0.1.1
+-----
+- nollieheel - Added frequency and mailto args for drupal cron job
+
 0.1.0
 -----
-- [nollieheel] - Initial release of chef-drupal
+- nollieheel - Initial release of chef-drupal